% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGT.HGTIndex.R
\name{HGT.HGTIndex}
\alias{HGT.HGTIndex}
\title{HGT.HGTIndex}
\usage{
HGT.HGTIndex(blast_output_donor, blast_output_recipient, best.bitscore = TRUE)
}
\arguments{
\item{blast_output_donor}{Input file with 12 columns in tab-delimited format}

\item{blast_output_recipient}{Input file with 12 columns in tab-delimited format}

\item{best.bitscore}{Logical. If TRUE, output includes best bitscores of donor and recipient.}
}
\value{
dataframe
}
\description{
'HGT.HGTIndex' calculates HGT Index (HI) for genes in target species from BLAST results between this species and donor/recipient species.
}
\details{
The output by BLAST or diamond BLAST will be used as input of 'HGT.HGTIndex' and should be in a tabular format with 12 columns. 

For BLAST, the output parameter should be: -outfmt 6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore. 

For diamond BLAST, the output parameter should be: -outfmt 6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore.
}
\examples{
test_donor <- read.table(gzfile(system.file("extdata", "test_0.txt.gz", package = "HGT.HGTIndex")), header=FALSE, sep="\t")
test_recipient <- read.table(gzfile(system.file("extdata", "test_1.txt.gz", package = "HGT.HGTIndex")), header=FALSE, sep="\t")
HGT.HGTIndex(test_donor, test_recipient, best.bitscore = TRUE)
}
\references{
1. L Yang, X Chen. Manuscript in preparation.
}
\author{
Lei Yang
}
